// models/settings.ts
export interface settings {
  frame: number;
  menu_music: number;
  note_burst: number;
  turntable: number;
  lane_cover: number;
  pacemaker_cover: number;
  note_skin: number;
  judge_font: number;
  note_beam: number;
  full_combo_splash: number;
  score_folders: boolean;
  clear_folders: boolean;
  difficulty_folders: boolean;
  alphabet_folders: boolean;
  hide_playcount: boolean;
  disable_graphcutin: boolean;
  classic_hispeed: boolean;
  hide_iidxid: boolean;
  disable_musicpreview: number;
  vefx_lock: number;
  effect: number;
  bomb_size: number;
  disable_hcn_color: number;
  first_note_preview: number;

  qpro_head: number;
  qpro_hair: number;
  qpro_hand: number;
  qpro_face: number;
  qpro_body: number;
}