// models/pc_data.ts
export interface pc_data {
  collection: "pc_data";

  version: number;

  sgid: number;
  dgid: number;

  deller: number;
  trophy: string[];

  sprank: number[];
  sppoint: number[];
  dprank: number[];
  dppoint: number[];

  spradar: number[];
  dpradar: number[];

  dp_clear_mission_clear: number;
  dp_clear_mission_level: number;
  dp_dj_mission_clear: number;
  dp_dj_mission_level: number;
  dp_level: number;
  dp_mission_point: number;
  dp_mplay: number;
  enemy_damage: number;
  progress: number;
  sp_clear_mission_clear: number;
  sp_clear_mission_level: number;
  sp_dj_mission_clear: number;
  sp_dj_mission_level: number;
  sp_level: number;
  sp_mission_point: number;
  sp_mplay: number;
  tips_read_list: number;

  dpnum: number;
  d_auto_scrach: number;
  d_camera_layout: number;
  d_disp_judge: number;
  d_gauge_disp: number;
  d_ghost_score: number;
  d_gno: number;
  d_graph_score: number;
  d_gtype: number;
  d_hispeed: number;
  d_judge: number;
  d_judgeAdj: number;
  d_lane_brignt: number;
  d_liflen: number;
  d_notes: number;
  d_opstyle: number;
  d_pace: number;
  d_sdlen: number;
  d_sdtype: number;
  d_sorttype: number;
  d_timing: number;
  d_tsujigiri_disp: number;
  dach: number;
  dp_opt: string;
  dp_opt2: string;
  d_exscore: number;
  d_sub_gno: number;
  d_tune: number;

  gpos: number;
  mode: number;
  pmode: number;
  rtype: number;
  ngrade: number;

  spnum: number;
  s_auto_scrach: number;
  s_camera_layout: number;
  s_disp_judge: number;
  s_gauge_disp: number;
  s_ghost_score: number;
  s_gno: number;
  s_graph_score: number;
  s_gtype: number;
  s_hispeed: number;
  s_judge: number;
  s_judgeAdj: number;
  s_lane_brignt: number;
  s_liflen: number;
  s_notes: number;
  s_opstyle: number;
  s_pace: number;
  s_sdlen: number;
  s_sdtype: number;
  s_sorttype: number;
  s_timing: number;
  s_tsujigiri_disp: number;
  sach: number;
  sp_opt: string;
  s_exscore: number;
  s_sub_gno: number;
  s_tune: number;
}

export const IIDX27_pc_data = {
  version: 27,

  deller: 0,
  sgid: -1,
  dgid: -1,
  trophy: ["0", "0", "0", "0", "0", "0", "0", "0", "0", "0"],
  sprank: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
  sppoint: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
  dprank: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
  dppoint: [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],

  spradar: [0, 0, 0, 0, 0, 0],
  dpradar: [0, 0, 0, 0, 0, 0],

  dp_clear_mission_clear: 0,
  dp_clear_mission_level: 0,
  dp_dj_mission_clear: 0,
  dp_dj_mission_level: 0,
  dp_level: 0,
  dp_mission_point: 0,
  dp_mplay: 0,
  enemy_damage: 0,
  progress: 0,
  sp_clear_mission_clear: 0,
  sp_clear_mission_level: 0,
  sp_dj_mission_clear: 0,
  sp_dj_mission_level: 0,
  sp_level: 0,
  sp_mission_point: 0,
  sp_mplay: 0,
  tips_read_list: 0,

  dpnum: 0,
  d_auto_scrach: 0,
  d_camera_layout: 0,
  d_disp_judge: 0,
  d_gauge_disp: 0,
  d_ghost_score: 0,
  d_gno: 0,
  d_graph_score: 0,
  d_gtype: 0,
  d_hispeed: 0,
  d_judge: 0,
  d_judgeAdj: 0,
  d_lane_brignt: 0,
  d_liflen: 0,
  d_notes: 0,
  d_opstyle: 0,
  d_pace: 0,
  d_sdlen: 0,
  d_sdtype: 0,
  d_sorttype: 0,
  d_timing: 0,
  d_tsujigiri_disp: 0,
  dach: 0,
  dp_opt: "0",
  dp_opt2: "0",
  d_sub_gno: 0,
  d_tune: 0,

  gpos: 0,
  mode: 0,
  pmode: 0,
  rtype: 0,

  spnum: 0,
  s_auto_scrach: 0,
  s_camera_layout: 0,
  s_disp_judge: 0,
  s_gauge_disp: 0,
  s_ghost_score: 0,
  s_gno: 0,
  s_graph_score: 0,
  s_gtype: 0,
  s_hispeed: 0,
  s_judge: 0,
  s_judgeAdj: 0,
  s_lane_brignt: 0,
  s_liflen: 0,
  s_notes: 0,
  s_opstyle: 0,
  s_pace: 0,
  s_sdlen: 0,
  s_sdtype: 0,
  s_sorttype: 0,
  s_timing: 0,
  s_tsujigiri_disp: 0,
  sach: 0,
  sp_opt: "0",
  s_sub_gno: 0,
  s_tune: 0,
};